pass org instead of array of orgs to OrgPreview
