+++ src/features/import/l10n/messageIds.ts
@@ -130,7 +130,7 @@ export default makeMessages('feat.import', {
         tags: m('Tags'),
       },
       next: m('Next'),
-      noOrgs: m('No organizations'),
+      noOrgs: m('No organization'),
       noTags: m('No tags'),
       noValue: m('No value'),
       previous: m('Previous'),
